requirements.txt
Flask==3.0.1
flask_cors==4.0.0
gunicorn==21.2.0
langchain==0.1.1
langchainhub==0.1.14
openai==1.8.0
pypdf==4.0.0
faiss_cpu==1.7.4
tiktoken==0.5.2
google-api-python-client==2.114.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0